typos and long -> integer

+++ dwsys/NUMfft_f.cpp
@@ -28,7 +28,7 @@
 #define FFT_DATA_TYPE float
 #include "NUMfft_core.h"
 
-void NUMforwardRealFastFourierTransform_f (float *data, long n) {
+void NUMforwardRealFastFourierTransform_f (float *data, integer n) {
 	struct NUMfft_Table_f table_struct;
 	NUMfft_Table_f table = &table_struct;
 
@@ -39,7 +39,7 @@ void NUMforwardRealFastFourierTransform_f (float *data, long n) {
 	if (n > 1) {
 		// To be compatible with old behaviour
 		float tmp = data[n];
-		for (long i = n; i > 2; i--) {
+		for (integer i = n; i > 2; i--) {
 			data[i] = data[i - 1];
 		}
 		data[2] = tmp;
@@ -48,14 +48,14 @@ void NUMforwardRealFastFourierTransform_f (float *data, long n) {
 	NUMfft_Table_free_f (table);
 }
 
-void NUMreverseRealFastFourierTransform_f (float *data, long n) {
+void NUMreverseRealFastFourierTransform_f (float *data, integer n) {
 	struct NUMfft_Table_f table_struct;
 	NUMfft_Table_f table = &table_struct;
 
 	if (n > 1) {
 		// To be compatible with old behaviour
 		float tmp = data[2];
-		for (long i = 2; i < n; i++) {
+		for (integer i = 2; i < n; i++) {
 			data[i] = data[i + 1];
 		}
 		data[n] = tmp;
@@ -80,10 +80,10 @@ void NUMfft_backward_f (NUMfft_Table_f me, float *data) {
 	drftb1 (my n, &data[1], my trigcache, my trigcache + my n, my splitcache);
 }
 
-void NUMfft_Table_init_f (NUMfft_Table_f me, long n) {
+void NUMfft_Table_init_f (NUMfft_Table_f me, integer n) {
 	my n = n;
 	my trigcache = NUMvector<float> (0, 3 * n - 1);
-	my splitcache = NUMvector<long> (0, 31);
+	my splitcache = NUMvector<integer> (0, 31);
 	NUMrffti (n, my trigcache, my splitcache);
 }
 
@@ -94,7 +94,7 @@ void NUMfft_Table_free_f (NUMfft_Table_f me) {
 	}
 }
 
-void NUMrealft_f (float *data, long n, int isign) {
+void NUMrealft_f (float *data, integer n, int isign) {
 	isign == 1 ? NUMforwardRealFastFourierTransform_f (data, n) :
 	NUMreverseRealFastFourierTransform_f (data, n);
 }
